Speed up detecting the NAND
It makes no sense to instruct the MTD framework to make use of the in flash
bad block table and then also add the flag to skip its search. This ends up
in scanning all page's OOB to get the bad block info again and again. And it
will slow down the start up process instead of shorten it (as the comment
suggests).

Signed-off-by: Juergen Beisert <jbe@pengutronix.de>